+++ src/hotspot/cpu/ppc/gc/z/zGlobals_ppc.cpp
@@ -107,7 +107,7 @@ static unsigned int probe_valid_max_address_bit(size_t init_bit, size_t min_bit)
   unsigned int max_valid_address_bit = 0;
   void* last_allocatable_address = nullptr;
 
-  const unsigned int page_size = os::vm_page_size();
+  const size_t page_size = os::vm_page_size();
 
   for (size_t i = init_bit; i >= min_bit; --i) {
     void* base_addr = (void*) (((unsigned long) 1U) << i);
