+++ src/hotspot/share/gc/shared/cardTable.cpp
@@ -99,8 +99,8 @@ void CardTable::initialize() {
 
   _cur_covered_regions = 0;
 
-  const size_t rs_align = _page_size == (size_t) os::vm_page_size() ? 0 :
-    MAX2(_page_size, (size_t) os::vm_allocation_granularity());
+  const size_t rs_align = _page_size == os::vm_page_size() ? 0 :
+    MAX2(_page_size, os::vm_allocation_granularity());
   ReservedSpace heap_rs(_byte_map_size, rs_align, _page_size);
 
   MemTracker::record_virtual_memory_type((address)heap_rs.base(), mtGC);
