+++ test/hotspot/gtest/memory/test_virtualspace.cpp
@@ -96,7 +96,7 @@ namespace {
   static void test_reserved_size_alignment_page_type(size_t size, size_t alignment, bool maybe_large) {
     if (size < alignment) {
       // Tests might set -XX:LargePageSizeInBytes=<small pages> and cause unexpected input arguments for this test.
-      ASSERT_EQ((size_t) os::vm_page_size(), os::large_page_size()) << "Test needs further refinement";
+      ASSERT_EQ(os::vm_page_size(), os::large_page_size()) << "Test needs further refinement";
       return;
     }
 
@@ -402,7 +402,7 @@ class TestReservedSpace : AllStatic {
   static void test_reserved_space3(size_t size, size_t alignment, bool maybe_large) {
     if (size < alignment) {
       // Tests might set -XX:LargePageSizeInBytes=<small pages> and cause unexpected input arguments for this test.
-      ASSERT_EQ((size_t)os::vm_page_size(), os::large_page_size()) << "Test needs further refinement";
+      ASSERT_EQ(os::vm_page_size(), os::large_page_size()) << "Test needs further refinement";
       return;
     }
